Revert "Merge remote-tracking branch 'origin/V-1.1' into V1.1-Jake"
This reverts commit b939c4f73bbb4f5cae01250b3eee5dad7f329656, reversing
changes made to 0bcc1fa38eb5172be326dacc2a392d1bc794d234.

+++ Engine/src/Architecture/AudioSystem/SoundManager.h
@@ -19,11 +19,11 @@ namespace IonixEngine
 
         bool LoadSound(const std::string& name, const std::string& filePath);
         Mix_Chunk* GetAudio(const std::string& name);
-        Mix_Chunk* GetAudio(uint32_t hash);
+        Mix_Chunk* GetAudio(uint64_t hash);
         void SetVolume(const std::string& name, float volume); // 0.0f to 1.0f
-        void SetVolume(uint32_t hash, float volume); // 0.0f to 1.0f
+        void SetVolume(uint64_t hash, float volume); // 0.0f to 1.0f
         float GetPlayTime(const std::string& alias);
-        float GetPlayTime(uint32_t hash);
+        float GetPlayTime(uint64_t hash);
 
     private:
         SoundManager() = default;
@@ -32,8 +32,8 @@ namespace IonixEngine
         SoundManager(const SoundManager&) = delete;
         SoundManager& operator=(const SoundManager&) = delete;
 
-        std::unordered_map<uint32_t, AudioData> m_Sounds;
-        std::unordered_map<uint32_t, float> m_Volumes;
+        std::unordered_map<uint64_t, AudioData> m_Sounds;
+        std::unordered_map<uint64_t, float> m_Volumes;
 
         SDL_AudioDeviceID m_Device = 0;
         SDL_AudioSpec m_DeviceSpec{};
