+++ Engine/src/Architecture/TextureManager/TextureManager.h
@@ -10,20 +10,21 @@ namespace IonixEngine
 	class TextureManager
 	{
 	private:
-		std::map<uint32_t, TextureData> textureDict;
+		std::map<uint64_t, TextureData> textureDict;
 		TextureData errorTexture;
 		SDL_Renderer* renderer;
 	public:
 		TextureManager();
 		void AddTexture(std::string filepath,std::string alias);
 		TextureData& GetTexture(std::string alias);
-		TextureData& GetTexture(uint32_t hash);
+		TextureData& GetTexture(uint64_t hash);
 
-		SDL_Texture* GetRawTexture(uint32_t hash);
+		SDL_Texture* GetRawTexture(uint64_t hash);
 
 		void Init();
 		void Shutdown();
-		static TextureManager& Get() {
+		static TextureManager& Get()
+		{
 			static TextureManager instance;
 			return instance;
 		};
